Updated, mordern, better looking textures for the planets. When hovering over then, they light up and the text on top of the screen shows the planet name. If no planet is selected, it says "Solar System"
Attempted to add orbits around the planets. Because of the close proximity, the Arc's look a little overcrowded/overwhelming. May want to consider putting it up as a costumisable option, or making it appear on planet select.

The last issue needed to be fixed is the radius line running across the midpoint of the arc and the edge. JavaFX doesn't have a "free hand" drawing, line-creating object, meaning the orbits are made from the outlines of an arc, which, even though it is a full 360 degrees, shows the middle radius line.